+++ provider.tf
@@ -19,6 +19,8 @@ terraform {
 provider "aws" {
   region = var.region
   default_tags {
-    tags = var.resource_tags
+    tags = merge({
+      Deployment = var.deployment_id
+    }, var.resource_tags)
   }
 }